exercise instructions
Exercise
Let's start off with a simple sentence:

Add a string to play with...

const sentence = "hello there from lighthouse labs";
Our goal is to animate the sentence, revealing one character at a time. This would make it look as though it is being typed in by someone, as shown in the GIF above.

Looping Through Characters
Let's loop through each character of the string and output it to the console.

Add the following code to our file and ensure that it runs as expected.

for (const char of sentence) {
  console.log(char);
}
The issue we see in the output is that each character is on a new line.

console.log will add a new line character (\n) to the end of the string. However, in Node we can use process.stdout.write instead to have more control and avoid automatically adding an extra "newline character" at the end each time.

Replace the console.log with process.stdout.write and confirm our results.

Indeed this prints the characters out on a single line. However, once the script completes, the next prompt appears on the same line, which is not ideal.

To stay focused, let's fix the prompt issue/bug at the very end. Although annoying, it's not blocking our progress at the moment.

Delaying The Display With setTimeout
We want the letters to appear with a delay, and setTimeout is all about delays. Let's use it here.

Add a setTimeout call within the loop, to delay the character output a bit.

setTimeout(() => {
  // print the char here
}, 1000) // <= 1s delay to make it noticeable. Can dial it down later.
Before running it, think about what the behaviour change will be, and why. It's important to see if our mental model of the control flow is accurate. Discuss with your peer / pair programmer prior to confirming.

Your Answer
timed out print will be displayed as per timer
Correct Answer
We'll see all the characters appear at once, one second after the program is run. There won't be any noticeable typing animation.

This is because, with each loop, we register a delayed output of each character, always by 1 second. The loop runs through the string quite fast, creating a 1s timer for each character. All the characters are set to appear after a 1s timeout.

For a three character string, for example, three 1s timeouts are created. One second later, all of them execute (in order).

Any micro-variances in delay, caused by the looping logic itself, will not be noticeable by us mere humans!

Delaying Each Character Separately
Instead of the characters all appearing at the same time later on, we need them each be staggered. Perhaps each one can be 50ms apart.

Implement the logic to have each letter delay its appearance by 50ms (or so) longer than the previous letter.

In other words, the timeout delays could like something like this:

Your implementation should not be tied to the example string. We should be able to change the string from "hello there from lighthouse labs" to any other string and still have the animation logic work.

Fixing the NewLine At The End
We still have one last thing to implement. It would be nice of the program to output a newline character before exiting, so that the next prompt starts on its own line.

Fix the bug we identified earlier by adding a newline character at the end of the animation.

Tip: We can use console.log, or write \n to stdout. That said, we can't output this character until all the characters are written out to stdout.